change many unsigned int types used in array indexing to size_t

+++ AmpTools/IUAmpTools/NormIntInterface.h
@@ -69,7 +69,7 @@ public:
   istream& loadNormIntCache( istream& in );
   void operator+=( const NormIntInterface& nii );
   
-  unsigned int numGenEvents() const { return m_nGenEvents; }
+  size_t numGenEvents() const { return m_nGenEvents; }
   double numAccEvents() const { return m_sumAccWeights; }
   
   // this integral folds in detector acceptance
@@ -101,7 +101,7 @@ public:
   const double* ampIntMatrix() const  { return m_ampIntCache;  }
   const double* normIntMatrix() const { return m_normIntCache; }
   
-  void setGenEvents( unsigned int events ) { m_nGenEvents = events; }
+  void setGenEvents( size_t events ) { m_nGenEvents = events; }
   void setAccEvents( double sumWeights ) { m_sumAccWeights = sumWeights; }
   
 protected:
@@ -135,7 +135,7 @@ private:
   mutable bool m_emptyNormIntCache;
   mutable bool m_emptyAmpIntCache;
   
-  mutable unsigned int m_nGenEvents;
+  mutable size_t m_nGenEvents;
   mutable double m_sumAccWeights;
   
 #ifndef __ACLIC__
